web_interface/backend/requirements.txt
pyrebase4
python-dotenv
flask
flask-cors